+++ src/components/formik/slider-control/index.tsx
